state: move proto aliases from api package
