+++ state/rollback.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"fmt"
 
-	cmtstate "github.com/cometbft/cometbft/api/cometbft/state"
 	cmtversion "github.com/cometbft/cometbft/api/cometbft/version/v1beta1"
+	cmtstate "github.com/cometbft/cometbft/state/proto"
 	"github.com/cometbft/cometbft/version"
 )
