+++ statesync/stateprovider.go
@@ -8,7 +8,6 @@ import (
 
 	dbm "github.com/cometbft/cometbft-db"
 
-	cmtstate "github.com/cometbft/cometbft/api/cometbft/state"
 	"github.com/cometbft/cometbft/libs/log"
 	cmtsync "github.com/cometbft/cometbft/libs/sync"
 	"github.com/cometbft/cometbft/light"
@@ -18,6 +17,7 @@ import (
 	lightdb "github.com/cometbft/cometbft/light/store/db"
 	rpchttp "github.com/cometbft/cometbft/rpc/client/http"
 	sm "github.com/cometbft/cometbft/state"
+	cmtstate "github.com/cometbft/cometbft/state/proto"
 	"github.com/cometbft/cometbft/types"
 	"github.com/cometbft/cometbft/version"
 )
